LayoutTests/platform/mac/imported/w3c/web-platform-tests/svg/path/error-handling/bounding-expected.txt

CONSOLE MESSAGE: line 10: Error: Problem parsing d="none"
CONSOLE MESSAGE: line 11: Error: Problem parsing d="# invalid"
CONSOLE MESSAGE: line 12: Error: Problem parsing d="m 0 0 l 3 -4 z # ignored suffix v 123"
CONSOLE MESSAGE: line 14: Error: Problem parsing d="m 0 0 l -9 11 -123 z # ignore last l parameter"
CONSOLE MESSAGE: line 96: TypeError: null is not an object (evaluating 'document.body.appendChild')
#PID UNRESPONSIVE - WebKitTestRunner (pid 24955)
FAIL: Timed out waiting for notifyDone to be called

#EOF
#EOF
